Adding ec2_alloc, which is a serverless way to manage an ec2 cluster (see allocate_ec2_instances docstring for more information). Some related changes: - run_one_job now tries to deallocate itself when it completes - Add launch_ec2_instance and launch_ec2_instances APIs, and make launch_ec2_instance async-ish so that you can launch more than one EC2 instance at a time - A bit of refactoring in aws_integration to use the new launch_ec2_instance API as well as a new ensure_security_group API. - Make _get_default_region_name work even if boto3.DEFAULT_SESSION hasn't been set yet

+++ src/meadowgrid/resource_allocation.py
@@ -37,7 +37,7 @@ class Resources:
     """
 
     memory_gb: float
-    logical_cpu: float
+    logical_cpu: int
     custom: Dict[str, float]
 
     def subtract(self, required: Resources) -> Optional[Resources]:
@@ -90,10 +90,10 @@ class Resources:
             memory_gb = 0.0
 
         if LOGICAL_CPU in resources_dict:
-            logical_cpu = resources_dict[LOGICAL_CPU]
+            logical_cpu = int(resources_dict[LOGICAL_CPU])
             del resources_dict[LOGICAL_CPU]
         else:
-            logical_cpu = 0.0
+            logical_cpu = 0
 
         return cls(memory_gb, logical_cpu, resources_dict)
